Safari-609.1.11.1/LayoutTests/js/nested-functions-expected.txt

This test ensures that properties of nested function code are not lost (due to cached function info).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS passed1 is true
PASS passed2 is true
PASS passed3 is true
PASS passed4 is true
PASS successfullyParsed is true

TEST COMPLETE
